feat: Implement histogram computation & project reorganization
- This commit intoduces a new way of handling histograms computations
- Integrates seamlessly with previously developed code
- Introduces major improvements in project organization
- Adds a separation of concerns & concepts
