--- deepfake_recognition/__init__.py
@@ -1,9 +1,0 @@
-from .models import DetectedFeatures
-from .types import NumpyArray
-from .helpers import detect_and_save_cropped_faces
-
-__all__ = [
-    "DetectedFeatures",
-    "NumpyArray",
-    "detect_and_save_cropped_faces",
-]
