+++ requirements.txt
@@ -1,0 +1,100 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile pyproject.toml -o requirements.txt
+annotated-types==0.7.0
+    # via pydantic
+certifi==2025.8.3
+    # via requests
+charset-normalizer==3.4.3
+    # via requests
+facenet-pytorch==2.5.3
+    # via deepfake-recognition (pyproject.toml)
+filelock==3.19.1
+    # via torch
+fsspec==2025.9.0
+    # via torch
+idna==3.10
+    # via requests
+jinja2==3.1.6
+    # via torch
+markupsafe==3.0.2
+    # via jinja2
+mpmath==1.3.0
+    # via sympy
+networkx==3.5
+    # via torch
+numpy==2.2.6
+    # via
+    #   facenet-pytorch
+    #   opencv-python
+    #   torchvision
+nvidia-cublas-cu12==12.8.4.1
+    # via
+    #   nvidia-cudnn-cu12
+    #   nvidia-cusolver-cu12
+    #   torch
+nvidia-cuda-cupti-cu12==12.8.90
+    # via torch
+nvidia-cuda-nvrtc-cu12==12.8.93
+    # via torch
+nvidia-cuda-runtime-cu12==12.8.90
+    # via torch
+nvidia-cudnn-cu12==9.10.2.21
+    # via torch
+nvidia-cufft-cu12==11.3.3.83
+    # via torch
+nvidia-cufile-cu12==1.13.1.3
+    # via torch
+nvidia-curand-cu12==10.3.9.90
+    # via torch
+nvidia-cusolver-cu12==11.7.3.90
+    # via torch
+nvidia-cusparse-cu12==12.5.8.93
+    # via
+    #   nvidia-cusolver-cu12
+    #   torch
+nvidia-cusparselt-cu12==0.7.1
+    # via torch
+nvidia-nccl-cu12==2.27.3
+    # via torch
+nvidia-nvjitlink-cu12==12.8.93
+    # via
+    #   nvidia-cufft-cu12
+    #   nvidia-cusolver-cu12
+    #   nvidia-cusparse-cu12
+    #   torch
+nvidia-nvtx-cu12==12.8.90
+    # via torch
+opencv-python==4.12.0.88
+    # via deepfake-recognition (pyproject.toml)
+pillow==11.3.0
+    # via
+    #   facenet-pytorch
+    #   torchvision
+pydantic==2.11.7
+    # via deepfake-recognition (pyproject.toml)
+pydantic-core==2.33.2
+    # via pydantic
+requests==2.32.5
+    # via facenet-pytorch
+setuptools==80.9.0
+    # via triton
+sympy==1.14.0
+    # via torch
+torch==2.8.0
+    # via torchvision
+torchvision==0.23.0
+    # via
+    #   deepfake-recognition (pyproject.toml)
+    #   facenet-pytorch
+triton==3.4.0
+    # via torch
+typing-extensions==4.15.0
+    # via
+    #   pydantic
+    #   pydantic-core
+    #   torch
+    #   typing-inspection
+typing-inspection==0.4.1
+    # via pydantic
+urllib3==2.5.0
+    # via requests
