+++ scripts/clean-pycache.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+# This script removes all __pycache__ directories and .pyc files from the current directory and its subdirectories.
+# Usage: ./clean-pycache.sh
+
+find . -type d -name "__pycache__" -exec rm -r {} +
